+++ extensions/vscode-aml/CHANGELOG.md
@@ -9,9 +9,10 @@ Check [Keep a Changelog](http://keepachangelog.com) for recommendations on how t
 ### Added
 
 - AML error reporting
-- AML rename schemas, entities and attributes
+- AML programmatic suggestions
+- AML rename: schemas, entities and attributes
 - AML conversion commands from/to SQL, JSON, DOT, Mermaid and more
-- Add "Open in Azimutt" link in preview
+- Add "Open in Azimutt" command and link in preview
 
 
 ## 0.1.1 - 2024-11-25
